Enable unixodbc so that Erlang's odbc compiles
Erlang/OTP's `odbc` library is optionally compiled. If a user's program
wants to use `odbc`, the current behavior is to use the host version of
`odbc`. This includes the shared libraries needed for `odbc` to be
included in system releases. It adds about 1 MB uncompressed to all
Nerves firmware using this system. The justification for this addition
is that `odbc` is part of Erlang/OTP so there's more of an expectation
of it working out-of-the-box.